+++ core/src/databases/remote_databases/salesforce/salesforce.rs
@@ -13,10 +13,6 @@ use tracing::error;
 
 use crate::cache;
 
-use crate::databases::remote_databases::salesforce::sandbox::{
-    convert::convert_to_soql, extract::extract_objects,
-};
-
 use crate::{
     databases::{
         database::{QueryDatabaseError, QueryResult, SqlDialect},
@@ -29,7 +25,7 @@ use crate::{
     },
 };
 
-use super::sandbox::structured_query::{StructuredQuery, Validator};
+use super::process_json_query::process_json_query;
 
 pub const MAX_QUERY_RESULT_ROWS: usize = 25_000;
 pub const GET_SESSION_MAX_TRIES: usize = 3;
@@ -647,21 +643,18 @@ impl RemoteDatabase for SalesforceRemoteDatabase {
         tables: &Vec<Table>,
         query: &str,
     ) -> Result<(Vec<QueryResult>, TableSchema, String), QueryDatabaseError> {
-        // Parse the JSON query
-        let parsed_query = serde_json::from_str::<StructuredQuery>(query).map_err(|e| {
-            QueryDatabaseError::ExecutionError(format!("Failed to parse JSON query: {}", e), None)
+        let processed_query = process_json_query(query).map_err(|e| {
+            QueryDatabaseError::ExecutionError(
+                format!(
+                    "Failed to process JSON query: {}",
+                    e.user_friendly_message()
+                ),
+                None,
+            )
         })?;
 
-        // Validate the structured query
-        if let Err(e) = parsed_query.validate() {
-            return Err(QueryDatabaseError::ExecutionError(
-                format!("Invalid structured query: {}", e),
-                None,
-            ));
-        }
-
         // Extract all objects referenced in the query
-        let referenced_objects = extract_objects(&parsed_query);
+        let referenced_objects = processed_query.objects;
 
         // Get the tables allowed in the query
         let allowed_tables: HashSet<String> = tables
@@ -679,22 +672,18 @@ impl RemoteDatabase for SalesforceRemoteDatabase {
         // Validate referenced objects against allowed tables using Salesforce metadata API
         // This will handle polymorphic relationships, plural forms, etc.
         if let Err(e) = self
-            .validate_referenced_objects(&referenced_objects, &allowed_tables, &parsed_query.object)
+            .validate_referenced_objects(
+                &referenced_objects,
+                &allowed_tables,
+                &processed_query.main_object,
+            )
             .await
         {
             return Err(e);
         }
 
-        // Convert the structured query to SOQL
-        let soql_query = convert_to_soql(&parsed_query).map_err(|e| {
-            QueryDatabaseError::ExecutionError(
-                format!("Error converting JSON query to SOQL: {}", e),
-                None,
-            )
-        })?;
-
         // Execute the SOQL query
-        self.execute_query(&soql_query).await
+        self.execute_query(&processed_query.soql).await
     }
 
     async fn get_tables_schema(&self, opaque_ids: &Vec<&str>) -> Result<Vec<TableSchema>> {
