+++ core/src/databases/remote_databases/salesforce/sandbox/validator.rs
@@ -1,663 +1,19 @@
-use super::error::SoqlError;
+use super::to_soql::ToSoql;
+use super::{
+    error::SoqlError,
+    models::{
+        Aggregate, FieldExpression, Filter, FunctionArgument, GroupBy, HavingClause, ParentField,
+        Relationship, StructuredQuery, TypedValue, WhereClause,
+    },
+};
 use lazy_static::lazy_static;
 use regex::Regex;
-use serde::{Deserialize, Serialize};
 
 const MAX_DOT_NOTATION_TRAVERSAL_DEPTH: usize = 1;
-const MAX_FUNCTION_NESTING_DEPTH: usize = 4; // Allow up to 4 levels of nesting
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct StructuredQuery {
-    /// The main Salesforce object to query (required).
-    pub object: String,
-
-    /// Fields to retrieve (optional if aggregates are present).
-    #[serde(default)]
-    pub fields: Vec<FieldExpression>,
-
-    /// Aggregate functions (optional if fields are present).
-    #[serde(default)]
-    pub aggregates: Vec<Aggregate>,
-
-    /// Filter conditions (optional).
-    #[serde(rename = "where")]
-    pub where_clause: Option<WhereClause>,
-
-    /// Sorting criteria (optional).
-    #[serde(default, rename = "orderBy")]
-    pub order_by: Vec<OrderBy>,
-
-    /// Result limit (optional).
-    pub limit: Option<u32>,
-
-    /// Result offset (optional).
-    pub offset: Option<u32>,
-
-    /// Parent-to-child relationships (optional).
-    #[serde(default)]
-    pub relationships: Vec<Relationship>,
-
-    /// Child-to-parent relationships (optional).
-    #[serde(default, rename = "parentFields")]
-    pub parent_fields: Vec<ParentField>,
-
-    /// Group by fields or structure (optional).
-    #[serde(rename = "groupBy")]
-    pub group_by: Option<GroupBy>,
-
-    /// Having clause for filtering aggregates (optional).
-    pub having: Option<HavingClause>,
-}
-
-/// Represents a where clause in a SOQL query.
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct WhereClause {
-    /// The logical condition for combining filters ("AND" or "OR").
-    #[serde(default = "default_logical_operator")]
-    pub condition: LogicalOperator,
-
-    /// The list of filters to apply.
-    pub filters: Vec<Filter>,
-}
-
-/// Default function for logical operator.
-fn default_logical_operator() -> LogicalOperator {
-    LogicalOperator::And
-}
-
-/// Represents a logical operator for combining filters.
-#[derive(Debug, Clone, Serialize, Deserialize)]
-#[serde(rename_all = "UPPERCASE")]
-pub enum LogicalOperator {
-    /// Logical AND operator.
-    And,
-
-    /// Logical OR operator.
-    Or,
-}
-
-/// Represents a strongly-typed value for use in SOQL queries
-#[derive(Debug, Clone, Serialize, Deserialize)]
-#[serde(untagged)]
-pub enum TypedValue {
-    /// A datetime value
-    DateTime {
-        #[serde(rename = "type")]
-        value_type: DateTimeType,
-        value: String,
-    },
-    /// A function call
-    Function {
-        #[serde(rename = "type")]
-        value_type: FunctionType,
-        function: String,
-        arguments: Vec<FunctionArgument>,
-    },
-    /// A regular JSON value
-    Regular(serde_json::Value),
-}
-
-/// Represents the type of a datetime value
-#[derive(Debug, Clone, Serialize, Deserialize)]
-#[serde(rename_all = "lowercase")]
-pub enum DateTimeType {
-    /// A datetime value
-    DateTime,
-}
-
-/// Represents the type of a function
-#[derive(Debug, Clone, Serialize, Deserialize)]
-#[serde(rename_all = "lowercase")]
-pub enum FunctionType {
-    /// A function call
-    Function,
-}
-
-/// Represents a function argument in a SOQL query.
-#[derive(Debug, Clone, Serialize, Deserialize)]
-#[serde(untagged)]
-pub enum FunctionArgument {
-    /// A field expression (field reference or function call)
-    Expression(FieldExpression),
-
-    /// A literal value
-    Literal(serde_json::Value),
-}
-
-/// Represents a field reference or function call in a SOQL query.
-#[derive(Debug, Clone, Serialize, Deserialize)]
-#[serde(untagged)]
-pub enum FieldExpression {
-    /// A simple field reference.
-    Field(String),
-
-    /// A function call.
-    Function {
-        function: String,
-        arguments: Vec<FunctionArgument>,
-    },
-}
-
-/// Represents a filter in a where clause.
-#[derive(Debug, Clone, Serialize, Deserialize)]
-#[serde(untagged)]
-pub enum Filter {
-    /// A simple field comparison.
-    Condition {
-        /// The field to filter on.
-        field: FieldExpression,
-
-        /// The comparison operator.
-        operator: String,
-
-        /// The value to compare against.
-        value: TypedValue,
-    },
-
-    /// A nested condition with its own logical operator and filters.
-    NestedCondition(WhereClause),
-}
-
-/// Represents an order by clause in a SOQL query.
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct OrderBy {
-    /// The field to sort by.
-    pub field: FieldExpression,
-
-    /// The sort direction (optional, defaults to "ASC").
-    #[serde(default = "default_order_direction")]
-    pub direction: OrderDirection,
-
-    /// Whether to put nulls first or last (optional).
-    pub nulls: Option<NullsPosition>,
-}
-
-/// Represents the sort direction in an order by clause.
-#[derive(Debug, Clone, Serialize, Deserialize)]
-#[serde(rename_all = "UPPERCASE")]
-pub enum OrderDirection {
-    /// Ascending order.
-    Asc,
-
-    /// Descending order.
-    Desc,
-}
-
-/// Represents the position of nulls in an order by clause.
-#[derive(Debug, Clone, Serialize, Deserialize)]
-#[serde(rename_all = "UPPERCASE")]
-pub enum NullsPosition {
-    /// Nulls first.
-    First,
-
-    /// Nulls last.
-    Last,
-}
-
-/// Represents a parent-to-child relationship in a SOQL query.
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct Relationship {
-    /// The name of the relationship.
-    #[serde(rename = "relationshipName")]
-    pub relationship_name: String,
-
-    /// The fields to retrieve from the related object.
-    pub fields: Vec<FieldExpression>,
-
-    /// Filter conditions for the related object (optional).
-    #[serde(rename = "where")]
-    pub where_clause: Option<WhereClause>,
-
-    /// Sorting criteria for the related object (optional).
-    #[serde(default, rename = "orderBy")]
-    pub order_by: Vec<OrderBy>,
-
-    /// Result limit for the related object (optional).
-    pub limit: Option<u32>,
-}
-
-/// Represents a child-to-parent relationship in a SOQL query.
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct ParentField {
-    /// The name of the relationship.
-    pub relationship: String,
-
-    /// The fields to retrieve from the parent object.
-    pub fields: Vec<FieldExpression>,
-}
-
-/// Represents an aggregate function in a SOQL query.
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct Aggregate {
-    /// The aggregate function to apply.
-    pub function: AggregateFunction,
-
-    /// The field to aggregate.
-    pub field: FieldExpression,
-
-    /// The alias for the aggregate result (optional).
-    pub alias: String,
-}
-
-/// Represents an aggregate function.
-#[derive(Debug, Clone, Serialize, Deserialize)]
-#[serde(rename_all = "UPPERCASE")]
-pub enum AggregateFunction {
-    /// Count function.
-    Count,
-
-    /// Sum function.
-    Sum,
-
-    /// Average function.
-    Avg,
-
-    /// Minimum function.
-    Min,
-
-    /// Maximum function.
-    Max,
-}
-
-/// Represents a group by clause in a SOQL query.
-#[derive(Debug, Clone, Serialize, Deserialize)]
-#[serde(untagged)]
-pub enum GroupBy {
-    /// A simple list of fields to group by.
-    Simple(Vec<FieldExpression>),
-
-    /// An advanced grouping structure (ROLLUP or CUBE).
-    Advanced {
-        /// The type of advanced grouping.
-        #[serde(rename = "type")]
-        group_type: GroupType,
-
-        /// The fields to group by.
-        fields: Vec<FieldExpression>,
-    },
-}
-
-/// Represents the type of advanced grouping.
-#[derive(Debug, Clone, Serialize, Deserialize)]
-#[serde(rename_all = "UPPERCASE")]
-pub enum GroupType {
-    /// ROLLUP grouping.
-    Rollup,
-
-    /// CUBE grouping.
-    Cube,
-}
-
-/// Represents a having clause in a SOQL query.
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct HavingClause {
-    /// The logical condition for combining filters ("AND" or "OR").
-    #[serde(default = "default_logical_operator")]
-    pub condition: LogicalOperator,
-
-    /// The list of aggregate filters to apply.
-    pub filters: Vec<AggregateFilter>,
-}
-
-/// Represents a filter in a having clause.
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct AggregateFilter {
-    /// The aggregate function to filter on.
-    pub function: AggregateFunction,
-
-    /// The field to aggregate.
-    pub field: String,
-
-    /// The comparison operator.
-    pub operator: String,
-
-    /// The value to compare against.
-    pub value: TypedValue,
-}
-
-/// Default function for order direction.
-fn default_order_direction() -> OrderDirection {
-    OrderDirection::Asc
-}
-
-// Checks that a path (dot notation) does not traverse more than `MAX_DOT_NOTATION_TRAVERSAL_DEPTH` level deep.
-fn validate_max_depth(path: &str, allowed_depth: usize) -> Result<(), SoqlError> {
-    let dot_count = path.chars().filter(|c| *c == '.').count();
-    if dot_count > allowed_depth {
-        return Err(SoqlError::field_validation_error(
-            path,
-            format!(
-                "Exceeds maximum dot notation depth of {} level(s)",
-                allowed_depth
-            ),
-        ));
-    }
-    Ok(())
-}
-
-fn is_valid_iso8601_datetime(value: &str) -> bool {
-    lazy_static! {
-        static ref ISO8601_REGEX: Regex =
-            Regex::new(r"^\d{4}-\d{2}-\d{2}(T\d{2}:\d{2}:\d{2}(\.\d+)?(Z|[+-]\d{2}:\d{2})?)?$")
-                .unwrap();
-    }
-
-    ISO8601_REGEX.is_match(value)
-}
-
-fn is_valid_date_literal(value: &str) -> bool {
-    // Check for simple literals
-    if matches!(
-        value,
-        "TODAY"
-            | "YESTERDAY"
-            | "TOMORROW"
-            | "THIS_WEEK"
-            | "LAST_WEEK"
-            | "NEXT_WEEK"
-            | "THIS_MONTH"
-            | "LAST_MONTH"
-            | "NEXT_MONTH"
-            | "THIS_QUARTER"
-            | "LAST_QUARTER"
-            | "NEXT_QUARTER"
-            | "THIS_YEAR"
-            | "LAST_YEAR"
-            | "NEXT_YEAR"
-            | "LAST_90_DAYS"
-            | "NEXT_90_DAYS"
-    ) {
-        return true;
-    }
-
-    // Check for parameterized literals with regex
-    lazy_static! {
-        static ref N_PATTERN: Regex =
-            Regex::new(r"^(LAST|NEXT)_N_(DAYS|WEEKS|MONTHS|QUARTERS|YEARS):\d+$").unwrap();
-        static ref FISCAL_PATTERN: Regex =
-            Regex::new(r"^(THIS|LAST|NEXT)_FISCAL_(QUARTER|YEAR)$").unwrap();
-        static ref FISCAL_N_PATTERN: Regex =
-            Regex::new(r"^(LAST|NEXT)_N_FISCAL_(QUARTERS|YEARS):\d+$").unwrap();
-    }
-
-    N_PATTERN.is_match(value) || FISCAL_PATTERN.is_match(value) || FISCAL_N_PATTERN.is_match(value)
-}
-
-fn is_valid_function_name(function: &str) -> bool {
-    matches!(
-        function,
-        // Date functions
-        "DAY_ONLY" | "CALENDAR_MONTH" | "CALENDAR_QUARTER" | "CALENDAR_YEAR" |
-        "DAY_IN_MONTH" | "DAY_IN_WEEK" | "DAY_IN_YEAR" |
-        "FISCAL_MONTH" | "FISCAL_QUARTER" | "FISCAL_YEAR" |
-        "HOUR_IN_DAY" | "WEEK_IN_MONTH" | "WEEK_IN_YEAR" |
-        // Formatting functions
-        "FORMAT" | "CONVERT_TIMEZONE" |
-        // Math functions
-        "ABS" | "CEILING" | "FLOOR" | "ROUND" |
-        // Geolocation functions
-        "DISTANCE" | "GEOLOCATION" |
-        // Logical functions
-        "CASE" | "NULLVALUE" | "ISBLANK" |
-        // Other supported functions
-        "CONVERTCURRENCY" | "TOLABEL"
-    )
-}
-
-/// Formats a JSON value for use in a SOQL query.
-fn format_json_value(value: &serde_json::Value) -> String {
-    match value {
-        serde_json::Value::Null => String::from("NULL"),
-        serde_json::Value::Bool(b) => b.to_string(),
-        serde_json::Value::Number(n) => n.to_string(),
-        serde_json::Value::String(s) => {
-            // Escape special characters in string values
-            // The main requirement is to escape single quotes by doubling them
-            let escaped = s.replace('\'', "''");
-            format!("'{}'", escaped)
-        }
-        serde_json::Value::Array(arr) => {
-            let values: Vec<String> = arr.iter().map(|v| format_json_value(v)).collect();
-            format!("({})", values.join(", "))
-        }
-        serde_json::Value::Object(_) => {
-            // Objects are not supported in SOQL values
-            String::from("NULL")
-        }
-    }
-}
-
-impl FunctionArgument {
-    pub fn format(&self) -> String {
-        match self {
-            FunctionArgument::Expression(expr) => expr.format(),
-            FunctionArgument::Literal(value) => format_json_value(value),
-        }
-    }
-}
-
-impl FieldExpression {
-    pub fn format(&self) -> String {
-        match self {
-            FieldExpression::Field(field) => field.clone(),
-            FieldExpression::Function {
-                function,
-                arguments,
-            } => {
-                let args_str = arguments
-                    .iter()
-                    .map(|arg| arg.format())
-                    .collect::<Vec<_>>()
-                    .join(", ");
-
-                format!("{}({})", function, args_str)
-            }
-        }
-    }
-}
-
-impl RecursiveValidator for FunctionArgument {
-    fn validate(&self, depth: usize) -> Result<(), SoqlError> {
-        // Check for maximum nesting depth
-        if depth > MAX_FUNCTION_NESTING_DEPTH {
-            return Err(SoqlError::field_validation_error(
-                "function",
-                format!(
-                    "Function nesting depth exceeds maximum of {}",
-                    MAX_FUNCTION_NESTING_DEPTH
-                ),
-            ));
-        }
-
-        match self {
-            FunctionArgument::Expression(expr) => {
-                match expr {
-                    FieldExpression::Field(field) => {
-                        // Field validation is already handled in FieldExpression::validate
-                        // but we need to ensure we don't hit the depth check again
-                        if field.is_empty() {
-                            return Err(SoqlError::field_validation_error(
-                                "<empty>",
-                                "Field name cannot be empty",
-                            ));
-                        }
-
-                        // Validate field dot notation depth
-                        validate_max_depth(field, MAX_DOT_NOTATION_TRAVERSAL_DEPTH)?;
-
-                        Ok(())
-                    }
-                    FieldExpression::Function {
-                        function,
-                        arguments,
-                    } => {
-                        // Validate function name
-                        if !is_valid_function_name(function) {
-                            return Err(SoqlError::field_validation_error(
-                                &format!("{}()", function),
-                                format!("Unsupported function: {}. Must be one of the supported Salesforce functions.", function),
-                            ));
-                        }
-
-                        // Validate that function has arguments
-                        if arguments.is_empty() {
-                            return Err(SoqlError::field_validation_error(
-                                &format!("{}()", function),
-                                format!("Function {} requires at least one argument", function),
-                            ));
-                        }
-
-                        // Validate each argument recursively with increased depth
-                        for (i, arg) in arguments.iter().enumerate() {
-                            if let Err(err) = arg.validate(depth + 1) {
-                                return Err(SoqlError::field_validation_error(
-                                    &format!("{}(...)", function),
-                                    format!(
-                                        "Argument {} for function {} is invalid: {}",
-                                        i + 1,
-                                        function,
-                                        err
-                                    ),
-                                ));
-                            }
-                        }
-
-                        // Check specific argument requirements for different functions
-                        match function.as_str() {
-                            // Date functions generally require exactly one date/datetime field
-                            "DAY_ONLY" | "CALENDAR_MONTH" | "CALENDAR_QUARTER"
-                            | "CALENDAR_YEAR" | "DAY_IN_MONTH" | "DAY_IN_WEEK" | "DAY_IN_YEAR"
-                            | "FISCAL_MONTH" | "FISCAL_QUARTER" | "FISCAL_YEAR" | "HOUR_IN_DAY"
-                            | "WEEK_IN_MONTH" | "WEEK_IN_YEAR" => {
-                                if arguments.len() != 1 {
-                                    return Err(SoqlError::field_validation_error(
-                                        &format!("{}(...)", function),
-                                        format!(
-                                            "{} function requires exactly one date/datetime field argument",
-                                            function
-                                        ),
-                                    ));
-                                }
-                            }
-                            // FORMAT function requires exactly two arguments (datetime field and format string)
-                            "FORMAT" => {
-                                if arguments.len() != 2 {
-                                    return Err(SoqlError::field_validation_error(
-                                        &format!("{}(...)", function),
-                                        "FORMAT function requires exactly two arguments: a date/datetime field and a format string",
-                                    ));
-                                }
-                            }
-                            // Other functions have varying requirements, add specific validations as needed
-                            _ => {}
-                        }
-
-                        Ok(())
-                    }
-                }
-            }
-            FunctionArgument::Literal(_) => {
-                // Literal values don't need special validation
-                Ok(())
-            }
-        }
-    }
-}
-
-impl Validator for FieldExpression {
-    fn validate(&self) -> Result<(), SoqlError> {
-        match self {
-            FieldExpression::Field(field) => {
-                // Check that field is not empty
-                if field.is_empty() {
-                    return Err(SoqlError::field_validation_error(
-                        "<empty>",
-                        "Field name cannot be empty",
-                    ));
-                }
-
-                // Validate field dot notation depth
-                validate_max_depth(field, MAX_DOT_NOTATION_TRAVERSAL_DEPTH)?;
-
-                Ok(())
-            }
-            FieldExpression::Function {
-                function,
-                arguments,
-            } => {
-                // Validate function name
-                if !is_valid_function_name(function) {
-                    return Err(SoqlError::field_validation_error(
-                        &format!("{}()", function),
-                        format!("Unsupported function: {}. Must be one of the supported Salesforce functions.", function),
-                    ));
-                }
-
-                // Validate that function has arguments
-                if arguments.is_empty() {
-                    return Err(SoqlError::field_validation_error(
-                        &format!("{}()", function),
-                        format!("Function {} requires at least one argument", function),
-                    ));
-                }
-
-                // Validate each argument with initial depth of 0
-                for (i, arg) in arguments.iter().enumerate() {
-                    if let Err(err) = RecursiveValidator::validate(arg, 0) {
-                        return Err(SoqlError::field_validation_error(
-                            &format!("{}(...)", function),
-                            format!(
-                                "Argument {} for function {} is invalid: {}",
-                                i + 1,
-                                function,
-                                err
-                            ),
-                        ));
-                    }
-                }
-
-                // Check specific argument requirements for different functions
-                match function.as_str() {
-                    // Date functions generally require exactly one date/datetime field
-                    "DAY_ONLY" | "CALENDAR_MONTH" | "CALENDAR_QUARTER" | "CALENDAR_YEAR"
-                    | "DAY_IN_MONTH" | "DAY_IN_WEEK" | "DAY_IN_YEAR" | "FISCAL_MONTH"
-                    | "FISCAL_QUARTER" | "FISCAL_YEAR" | "HOUR_IN_DAY" | "WEEK_IN_MONTH"
-                    | "WEEK_IN_YEAR" => {
-                        if arguments.len() != 1 {
-                            return Err(SoqlError::field_validation_error(
-                                &format!("{}(...)", function),
-                                format!(
-                                    "{} function requires exactly one date/datetime field argument",
-                                    function
-                                ),
-                            ));
-                        }
-                    }
-                    // FORMAT function requires exactly two arguments
-                    "FORMAT" => {
-                        if arguments.len() != 2 {
-                            return Err(SoqlError::field_validation_error(
-                                &format!("{}(...)", function),
-                                "FORMAT function requires exactly two arguments: a date/datetime field and a format string",
-                            ));
-                        }
-                    }
-                    // Other functions have varying requirements, add specific validations as needed
-                    _ => {}
-                }
-
-                Ok(())
-            }
-        }
-    }
-}
+const MAX_FUNCTION_NESTING_DEPTH: usize = 3; // Allow up to 3 levels of nesting
 
 pub trait Validator {
     fn validate(&self) -> Result<(), SoqlError>;
-}
-
-pub trait RecursiveValidator {
-    fn validate(&self, depth: usize) -> Result<(), SoqlError>;
 }
 
 impl Validator for StructuredQuery {
@@ -737,6 +93,114 @@ impl Validator for StructuredQuery {
     }
 }
 
+impl FunctionArgument {
+    fn validate_recursive(&self, depth: usize) -> Result<(), SoqlError> {
+        // Check for maximum nesting depth
+        if depth > MAX_FUNCTION_NESTING_DEPTH {
+            return Err(SoqlError::field_validation_error(
+                "function",
+                format!(
+                    "Function nesting depth exceeds maximum of {}",
+                    MAX_FUNCTION_NESTING_DEPTH
+                ),
+            ));
+        }
+
+        match self {
+            FunctionArgument::Expression(expr) => {
+                match expr {
+                    FieldExpression::Field(field) => {
+                        // Field validation is already handled in FieldExpression::validate
+                        // but we need to ensure we don't hit the depth check again
+                        if field.is_empty() {
+                            return Err(SoqlError::field_validation_error(
+                                "<empty>",
+                                "Field name cannot be empty",
+                            ));
+                        }
+
+                        // Validate field dot notation depth
+                        validate_max_depth(field, MAX_DOT_NOTATION_TRAVERSAL_DEPTH)?;
+
+                        Ok(())
+                    }
+                    FieldExpression::Function {
+                        function,
+                        arguments,
+                    } => {
+                        // Validate function name
+                        if !is_valid_function_name(function) {
+                            return Err(SoqlError::field_validation_error(
+                                &format!("{}()", function),
+                                format!("Unsupported function: {}. Must be one of the supported Salesforce functions.", function),
+                            ));
+                        }
+
+                        // Validate that function has arguments
+                        if arguments.is_empty() {
+                            return Err(SoqlError::field_validation_error(
+                                &format!("{}()", function),
+                                format!("Function {} requires at least one argument", function),
+                            ));
+                        }
+
+                        // Validate each argument recursively with increased depth
+                        for (i, arg) in arguments.iter().enumerate() {
+                            if let Err(err) = arg.validate_recursive(depth + 1) {
+                                return Err(SoqlError::field_validation_error(
+                                    &format!("{}(...)", function),
+                                    format!(
+                                        "Argument {} for function {} is invalid: {}",
+                                        i + 1,
+                                        function,
+                                        err
+                                    ),
+                                ));
+                            }
+                        }
+
+                        // Check specific argument requirements for different functions
+                        match function.as_str() {
+                            // Date functions generally require exactly one date/datetime field
+                            "DAY_ONLY" | "CALENDAR_MONTH" | "CALENDAR_QUARTER"
+                            | "CALENDAR_YEAR" | "DAY_IN_MONTH" | "DAY_IN_WEEK" | "DAY_IN_YEAR"
+                            | "FISCAL_MONTH" | "FISCAL_QUARTER" | "FISCAL_YEAR" | "HOUR_IN_DAY"
+                            | "WEEK_IN_MONTH" | "WEEK_IN_YEAR" => {
+                                if arguments.len() != 1 {
+                                    return Err(SoqlError::field_validation_error(
+                                        &format!("{}(...)", function),
+                                        format!(
+                                            "{} function requires exactly one date/datetime field argument",
+                                            function
+                                        ),
+                                    ));
+                                }
+                            }
+                            // FORMAT function requires exactly two arguments (datetime field and format string)
+                            "FORMAT" => {
+                                if arguments.len() != 2 {
+                                    return Err(SoqlError::field_validation_error(
+                                        &format!("{}(...)", function),
+                                        "FORMAT function requires exactly two arguments: a date/datetime field and a format string",
+                                    ));
+                                }
+                            }
+                            // Other functions have varying requirements, add specific validations as needed
+                            _ => {}
+                        }
+
+                        Ok(())
+                    }
+                }
+            }
+            FunctionArgument::Literal(_) => {
+                // Literal values don't need special validation
+                Ok(())
+            }
+        }
+    }
+}
+
 impl Validator for TypedValue {
     fn validate(&self) -> Result<(), SoqlError> {
         match self {
@@ -778,7 +242,7 @@ impl Validator for TypedValue {
 
                 // Validate each argument
                 for (i, arg) in arguments.iter().enumerate() {
-                    if let Err(err) = RecursiveValidator::validate(arg, 0) {
+                    if let Err(err) = arg.validate_recursive(0) {
                         return Err(SoqlError::value_error(
                             "function",
                             format!(
@@ -848,7 +312,7 @@ impl Validator for WhereClause {
                     // Validate value
                     value.validate()?;
 
-                    let field_name = field.format();
+                    let field_name = field.to_soql();
 
                     // Get the underlying value for operator validation
                     let json_value = match value {
@@ -977,6 +441,14 @@ impl Validator for ParentField {
         // Validate field expressions
         for field in &self.fields {
             field.validate()?;
+            match field {
+                FieldExpression::Field(_) => {}
+                FieldExpression::Function { .. } => {
+                    return Err(SoqlError::unsupported_feature(
+                        "Function calls are not supported on parent fields",
+                    ));
+                }
+            }
         }
 
         Ok(())
@@ -1035,10 +507,10 @@ impl Validator for HavingClause {
 impl Validator for Aggregate {
     fn validate(&self) -> Result<(), SoqlError> {
         self.field.validate()?;
-        let field_str = self.field.format();
+        let field_str = self.field.to_soql();
         if self.alias.is_empty() {
             return Err(SoqlError::field_validation_error(
-                format!("{}({})", self.function.as_str(), field_str),
+                format!("{}({})", self.function.to_soql(), field_str),
                 "Aggregate function must have a non-empty alias",
             ));
         }
@@ -1047,101 +519,185 @@ impl Validator for Aggregate {
     }
 }
 
-// Helper method to get string representation of AggregateFunction
-impl AggregateFunction {
-    pub fn as_str(&self) -> &'static str {
+impl Validator for FieldExpression {
+    fn validate(&self) -> Result<(), SoqlError> {
         match self {
-            AggregateFunction::Count => "COUNT",
-            AggregateFunction::Sum => "SUM",
-            AggregateFunction::Avg => "AVG",
-            AggregateFunction::Min => "MIN",
-            AggregateFunction::Max => "MAX",
-        }
-    }
-}
+            FieldExpression::Field(field) => {
+                // Check that field is not empty
+                if field.is_empty() {
+                    return Err(SoqlError::field_validation_error(
+                        "<empty>",
+                        "Field name cannot be empty",
+                    ));
+                }
 
-// Helper method to get string representation of GroupType
-impl GroupType {
-    pub fn as_str(&self) -> &'static str {
-        match self {
-            GroupType::Rollup => "ROLLUP",
-            GroupType::Cube => "CUBE",
-        }
-    }
-}
+                // Validate field dot notation depth
+                validate_max_depth(field, MAX_DOT_NOTATION_TRAVERSAL_DEPTH)?;
 
-impl TypedValue {
-    pub fn format(&self) -> String {
-        match self {
-            TypedValue::DateTime { value, .. } => {
-                // For datetime objects, just return the value without quotes
-                value.clone()
+                Ok(())
             }
-            TypedValue::Function {
+            FieldExpression::Function {
                 function,
                 arguments,
-                ..
             } => {
-                // Format function call: FUNCTION(arg1, arg2, ...)
-                let args_str = arguments
-                    .iter()
-                    .map(|arg| arg.format())
-                    .collect::<Vec<_>>()
-                    .join(", ");
+                // Validate function name
+                if !is_valid_function_name(function) {
+                    return Err(SoqlError::field_validation_error(
+                        &format!("{}()", function),
+                        format!("Unsupported function: {}. Must be one of the supported Salesforce functions.", function),
+                    ));
+                }
 
-                format!("{}({})", function, args_str)
+                // Validate that function has arguments
+                if arguments.is_empty() {
+                    return Err(SoqlError::field_validation_error(
+                        &format!("{}()", function),
+                        format!("Function {} requires at least one argument", function),
+                    ));
+                }
+
+                // Validate each argument with initial depth of 1
+                for (i, arg) in arguments.iter().enumerate() {
+                    if let Err(err) = arg.validate_recursive(1) {
+                        return Err(SoqlError::field_validation_error(
+                            &format!("{}(...)", function),
+                            format!(
+                                "Argument {} for function {} is invalid: {}",
+                                i + 1,
+                                function,
+                                err
+                            ),
+                        ));
+                    }
+                }
+
+                // Check specific argument requirements for different functions
+                match function.as_str() {
+                    // Date functions generally require exactly one date/datetime field
+                    "DAY_ONLY" | "CALENDAR_MONTH" | "CALENDAR_QUARTER" | "CALENDAR_YEAR"
+                    | "DAY_IN_MONTH" | "DAY_IN_WEEK" | "DAY_IN_YEAR" | "FISCAL_MONTH"
+                    | "FISCAL_QUARTER" | "FISCAL_YEAR" | "HOUR_IN_DAY" | "WEEK_IN_MONTH"
+                    | "WEEK_IN_YEAR" => {
+                        if arguments.len() != 1 {
+                            return Err(SoqlError::field_validation_error(
+                                &format!("{}(...)", function),
+                                format!(
+                                    "{} function requires exactly one date/datetime field argument",
+                                    function
+                                ),
+                            ));
+                        }
+                    }
+                    // FORMAT function requires exactly two arguments
+                    "FORMAT" => {
+                        if arguments.len() != 2 {
+                            return Err(SoqlError::field_validation_error(
+                                &format!("{}(...)", function),
+                                "FORMAT function requires exactly two arguments: a date/datetime field and a format string",
+                            ));
+                        }
+                    }
+                    // Other functions have varying requirements, add specific validations as needed
+                    _ => {}
+                }
+
+                Ok(())
             }
-            TypedValue::Regular(json_value) => format_json_value(json_value),
         }
     }
+}
+
+// Checks that a path (dot notation) does not traverse more than `MAX_DOT_NOTATION_TRAVERSAL_DEPTH` level deep.
+fn validate_max_depth(path: &str, allowed_depth: usize) -> Result<(), SoqlError> {
+    let dot_count = path.chars().filter(|c| *c == '.').count();
+    if dot_count > allowed_depth {
+        return Err(SoqlError::field_validation_error(
+            path,
+            format!(
+                "Exceeds maximum dot notation depth of {} level(s)",
+                allowed_depth
+            ),
+        ));
+    }
+    Ok(())
+}
+
+fn is_valid_iso8601_datetime(value: &str) -> bool {
+    lazy_static! {
+        static ref ISO8601_REGEX: Regex =
+            Regex::new(r"^\d{4}-\d{2}-\d{2}(T\d{2}:\d{2}:\d{2}(\.\d+)?(Z|[+-]\d{2}:\d{2})?)?$")
+                .unwrap();
+    }
+
+    ISO8601_REGEX.is_match(value)
+}
+
+fn is_valid_date_literal(value: &str) -> bool {
+    // Check for simple literals
+    if matches!(
+        value,
+        "TODAY"
+            | "YESTERDAY"
+            | "TOMORROW"
+            | "THIS_WEEK"
+            | "LAST_WEEK"
+            | "NEXT_WEEK"
+            | "THIS_MONTH"
+            | "LAST_MONTH"
+            | "NEXT_MONTH"
+            | "THIS_QUARTER"
+            | "LAST_QUARTER"
+            | "NEXT_QUARTER"
+            | "THIS_YEAR"
+            | "LAST_YEAR"
+            | "NEXT_YEAR"
+            | "LAST_90_DAYS"
+            | "NEXT_90_DAYS"
+    ) {
+        return true;
+    }
+
+    // Check for parameterized literals with regex
+    lazy_static! {
+        static ref N_PATTERN: Regex =
+            Regex::new(r"^(LAST|NEXT)_N_(DAYS|WEEKS|MONTHS|QUARTERS|YEARS):\d+$").unwrap();
+        static ref FISCAL_PATTERN: Regex =
+            Regex::new(r"^(THIS|LAST|NEXT)_FISCAL_(QUARTER|YEAR)$").unwrap();
+        static ref FISCAL_N_PATTERN: Regex =
+            Regex::new(r"^(LAST|NEXT)_N_FISCAL_(QUARTERS|YEARS):\d+$").unwrap();
+    }
+
+    N_PATTERN.is_match(value) || FISCAL_PATTERN.is_match(value) || FISCAL_N_PATTERN.is_match(value)
+}
+
+fn is_valid_function_name(function: &str) -> bool {
+    matches!(
+        function,
+        // Date functions
+        "DAY_ONLY" | "CALENDAR_MONTH" | "CALENDAR_QUARTER" | "CALENDAR_YEAR" |
+        "DAY_IN_MONTH" | "DAY_IN_WEEK" | "DAY_IN_YEAR" |
+        "FISCAL_MONTH" | "FISCAL_QUARTER" | "FISCAL_YEAR" |
+        "HOUR_IN_DAY" | "WEEK_IN_MONTH" | "WEEK_IN_YEAR" |
+        // Formatting functions
+        "FORMAT" | "CONVERT_TIMEZONE" |
+        // Math functions
+        "ABS" | "CEILING" | "FLOOR" | "ROUND" |
+        // Geolocation functions
+        "DISTANCE" | "GEOLOCATION" |
+        // Logical functions
+        "CASE" | "NULLVALUE" | "ISBLANK" |
+        // Other supported functions
+        "CONVERTCURRENCY" | "TOLABEL"
+    )
 }
 
 #[cfg(test)]
 mod tests {
+    use crate::databases::remote_databases::salesforce::sandbox::models::{
+        AggregateFilter, AggregateFunction, DateTimeType, LogicalOperator,
+    };
+
     use super::*;
-
-    #[test]
-    fn test_format_json_value_escaping() {
-        // Test null value
-        assert_eq!(format_json_value(&serde_json::Value::Null), "NULL");
-
-        // Test boolean value
-        assert_eq!(format_json_value(&serde_json::json!(true)), "true");
-        assert_eq!(format_json_value(&serde_json::json!(false)), "false");
-
-        // Test number value
-        assert_eq!(format_json_value(&serde_json::json!(42)), "42");
-        assert_eq!(format_json_value(&serde_json::json!(3.14)), "3.14");
-
-        // Test string value with special characters
-        assert_eq!(format_json_value(&serde_json::json!("hello")), "'hello'");
-        assert_eq!(
-            format_json_value(&serde_json::json!("O'Reilly")),
-            "'O''Reilly'"
-        ); // Single quotes are doubled
-        assert_eq!(
-            format_json_value(&serde_json::json!("Line1\nLine2")),
-            "'Line1\nLine2'"
-        ); // Newlines preserved
-
-        // Test array value
-        assert_eq!(
-            format_json_value(&serde_json::json!(["a", "b", "c"])),
-            "('a', 'b', 'c')"
-        );
-
-        // Test mixed array
-        assert_eq!(
-            format_json_value(&serde_json::json!(["a", 1, true, null])),
-            "('a', 1, true, NULL)"
-        );
-
-        // Test object (not supported)
-        assert_eq!(
-            format_json_value(&serde_json::json!({"key": "value"})),
-            "NULL"
-        );
-    }
 
     #[test]
     fn test_date_literals_validation() {
@@ -1663,7 +1219,7 @@ mod tests {
 
         assert!(
             nested_function.validate().is_ok(),
-            "Nested functions should now be allowed"
+            "Nested functions should be allowed"
         );
 
         // Invalid field expression with deep dot notation
@@ -1700,22 +1256,16 @@ mod tests {
 
         let result = deep_nesting.validate();
 
-        // Our current depth is 4 (FORMAT -> CALENDAR_MONTH -> CALENDAR_YEAR -> DAY_ONLY)
-        // If MAX_FUNCTION_NESTING_DEPTH is >= 4, this will pass, otherwise it will fail
-        if MAX_FUNCTION_NESTING_DEPTH >= 4 {
-            assert!(result.is_ok(), "Expected valid nesting depth to pass");
-        } else {
-            assert!(result.is_err(), "Excessive nesting should not be allowed");
+        assert!(result.is_err(), "Excessive nesting should not be allowed");
 
-            if let Err(SoqlError::FieldValidationError { reason, .. }) = result {
-                assert!(
-                    reason.contains("Function nesting depth exceeds maximum"),
-                    "Error message should mention that maximum nesting depth is exceeded: {}",
-                    reason
-                );
-            } else {
-                panic!("Expected a field validation error about maximum nesting depth");
-            }
+        if let Err(SoqlError::FieldValidationError { reason, .. }) = result {
+            assert!(
+                reason.contains("Function nesting depth exceeds maximum"),
+                "Error message should mention that maximum nesting depth is exceeded: {}",
+                reason
+            );
+        } else {
+            panic!("Expected a field validation error about maximum nesting depth");
         }
     }
 }
